Problema resolvido! Agora tudo esta rodando!

+++ declaracoes.hpp
@@ -1,6 +1,6 @@
 #include <iostream>
 
 void trataInput(char inputName[], char tipoEntrada[]);
-void primalDual(int qtdCli, int qtdInst, float * custoF, float * custoA);
+void primalDual(int qtdCli, int qtdInst, double * custoF, double * custoA);
 
 void teste();
